requirements.txt
airportsdata==20250909
beautifulsoup4==4.14.3
Flask==3.1.3
FlightRadarAPI==1.4.0
geocoder==1.38.1
python-dotenv==1.2.1
requests==2.32.5
